moving pinata, proxy icon file to utils
