+++ components/rmrk/Create/CreateCollection.vue
@@ -105,7 +105,7 @@ import SubscribeMixin from '@/utils/mixins/subscribeMixin'
 import RmrkVersionMixin from '@/utils/mixins/rmrkVersionMixin'
 import { Collection, CollectionMetadata } from '../service/scheme'
 import { unSanitizeIpfsUrl } from '@/utils/ipfs'
-import { pinJson, pinFileDirect } from '@/proxy'
+import { pinJson, pinFileDirect } from '@/utils/proxy'
 import { decodeAddress } from '@polkadot/keyring'
 import { u8aToHex } from '@polkadot/util'
 import { generateId } from '@/components/rmrk/service/Consolidator'
