+++ components/rmrk/Create/SimpleMint.vue
@@ -236,7 +236,7 @@ import {
   getNftId,
 } from '../service/scheme'
 import { unSanitizeIpfsUrl } from '@/utils/ipfs'
-import { pinJson, getKey, revokeKey } from '@/proxy'
+import { pinJson, getKey, revokeKey } from '@/utils/proxy'
 import { formatBalance } from '@polkadot/util'
 import { generateId } from '@/components/rmrk/service/Consolidator'
 import { supportTx, calculateCost, offsetTx, feeTx } from '@/utils/support'
@@ -244,7 +244,7 @@ import { resolveMedia } from '../utils'
 import NFTUtils, { MintType } from '../service/NftUtils'
 import { DispatchError } from '@polkadot/types/interfaces'
 import { ipfsToArweave } from '@/utils/ipfs'
-import { APIKeys, pinFile as pinFileToIPFS } from '@/pinata'
+import { APIKeys, pinFile as pinFileToIPFS } from '@/utils/pinata'
 import TransactionMixin from '@/utils/mixins/txMixin'
 import { encodeAddress, isAddress } from '@polkadot/util-crypto'
 import ChainMixin from '@/utils/mixins/chainMixin'
