+++ components/rmrk/Gallery/AvailableActions.vue
@@ -41,7 +41,7 @@ import { Component, mixins, Prop} from 'nuxt-property-decorator'
 import Connector from '@vue-polkadot/vue-api'
 import exec, { execResultValue, txCb } from '@/utils/transactionExecutor'
 import { notificationTypes, showNotification } from '@/utils/notification'
-import { unpin } from '@/proxy'
+import { unpin } from '@/utils/proxy'
 import RmrkVersionMixin from '@/utils/mixins/rmrkVersionMixin'
 import { somePercentFromTX } from '@/utils/support'
 import shouldUpdate from '@/utils/shouldUpdate'
