+++ utils/pinata.ts
@@ -1,5 +1,5 @@
 import Axios from 'axios'
-import { extractCid, justHash } from './utils/ipfs'
+import { extractCid, justHash } from './ipfs'
 
 export const BASE_URL = 'https://api.pinata.cloud/pinning/'
